(wip) reorg 3d place code
