+++ include/hydra/places/gvd_places/graph_extractor.h
@@ -37,10 +37,11 @@
 
 #include <queue>
 
-#include "hydra/places/graph_extractor_utilities.h"
-#include "hydra/places/gvd_graph.h"
-#include "hydra/places/gvd_merge_policies.h"
-#include "hydra/places/gvd_voxel.h"
+#include "hydra/places/gvd_places/compressed_node.h"
+#include "hydra/places/gvd_places/graph_extractor_utilities.h"
+#include "hydra/places/gvd_places/gvd_graph.h"
+#include "hydra/places/gvd_places/gvd_merge_policies.h"
+#include "hydra/places/gvd_places/gvd_voxel.h"
 #include "hydra/reconstruction/voxel_types.h"
 
 namespace hydra::places {
@@ -50,32 +51,6 @@ struct GvdParentTracker;
 struct IndexVoxelPair {
   GlobalIndex index;
   const GvdVoxel* voxel;
-};
-
-struct CompressedNode {
-  using CompressedNodeMap = std::unordered_map<uint64_t, CompressedNode>;
-  uint64_t node_id;
-  std::set<uint64_t> siblings;
-  std::set<uint64_t> active_refs;
-  std::set<uint64_t> archived_refs;
-  std::unordered_map<uint64_t, std::map<uint64_t, uint64_t>> sibling_support;
-  std::map<uint64_t, uint64_t> sibling_ref_counts;
-  uint64_t best_gvd_id;
-  bool in_graph = false;
-
-  explicit CompressedNode(uint64_t node_id);
-
-  void addEdgeObservation(uint64_t gvd_id,
-                          uint64_t neighbor_gvd_id,
-                          uint64_t sibling_id);
-
-  bool removeEdgeObservation(uint64_t gvd_id, uint64_t neighbor_gvd_id);
-
-  std::list<uint64_t> removeEdgeObservations(uint64_t gvd_id, CompressedNodeMap& nodes);
-
-  void mergeObservations(uint64_t original_id, uint64_t new_id);
-
-  void merge(CompressedNode& other, CompressedNodeMap& nodes);
 };
 
 using IndexVoxelQueue = std::list<IndexVoxelPair>;
