+++ include/hydra/utils/layer_io.h
@@ -43,7 +43,7 @@
 #include <typeinfo>
 #include <vector>
 
-#include "hydra/places/gvd_voxel.h"
+#include "hydra/places/gvd_places/gvd_voxel.h"
 #include "hydra/reconstruction/voxel_types.h"
 
 namespace hydra::io {
